Move Rust tests to their respective files
